Wynter/requirements.txt
aiodns==2.0.0
aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
cchardet==2.1.7
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
discord.py==1.5.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycares==3.1.1
pycparser==2.20
pylint==2.6.0
PyMySQL==0.10.1
python-dotenv==0.15.0
requests==2.25.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
yarl==1.5.1
